% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{all_combn}
\alias{all_combn}
\title{Compute all permutations (with repeated draws)}
\usage{
all_combn(vec_, n = 1)
}
\arguments{
\item{vec_}{A vector to draw from .}

\item{n}{default to 1, is the sample size.}
}
\value{
a matrix with vec_^n rows and n columns.
}
\description{
Compute all permutations (with repeated draws)
}
\examples{
all_combn(c(1:3), 2)

}
