% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{all_insurance}
\alias{all_insurance}
\title{Compute all possible sets of insured individuals among eligible members}
\usage{
all_insurance(vec_)
}
\arguments{
\item{vec_}{A vector of indices of eligible members.}
}
\value{
combinations of all possible insurance bundles. Each is a vector of indices of members with insurance
}
\description{
Compute all possible sets of insured individuals among eligible members
}
\examples{
all_insurance(c(1,2,3))

}
