% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{counterfactual_household_draw_theta_kappa_Rdraw}
\alias{counterfactual_household_draw_theta_kappa_Rdraw}
\title{Compute the counterfactual draws (preference parameters independent). This is different from household_draw_theta_kappa_Rdraw because the draws of thetabar
is only conducted once, and there is no integration over thetabar. Also, the insurance bundles are not one-person deviation from the observed bundle, but
all possible bundles.}
\usage{
counterfactual_household_draw_theta_kappa_Rdraw(
  hh_index,
  param,
  n_draw_halton = 1000,
  n_draw_gauss = 10,
  sick_parameters,
  xi_parameters,
  u_lowerbar = -10,
  policy_mat_hh,
  seed_number = 1,
  constraint_function,
  within_hh_heterogeneity = list(omega = TRUE, gamma = TRUE, delta = TRUE, theta_bar =
    TRUE),
  income_vec = NA
)
}
\arguments{
\item{hh_index}{index (from object data_hh_list) of a particular household}

\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}

\item{n_draw_halton}{this is the number of halton draws, default to 1000}

\item{n_draw_gauss}{this is number of Gauss  points.}

\item{sick_parameters}{estimated sick_parameters}

\item{xi_parameters}{estimated xi_parameters}

\item{u_lowerbar}{value of u when income is negative}

\item{policy_mat_hh}{is the policy matrix (might be actual or a counterfactual), should follow the format of policy_mat[\link{hh_index}]}

\item{constraint_function}{is a function that can reflect whether we are imposing pure bundling or something else on the possible choice set. For example, for pure bundling, constraint_function = function(x) {x\link{-c(1,length(x))} = -Inf; return(x)}}

\item{within_hh_heterogeneity}{is a list with list names gamma, delta, and theta_bar. If a list element is true, within-hh-heterogeneity is allowed.}

\item{income_vec}{a vector of income net premium at different number of voluntarily insured members (from 0 to HHsize_s)}
}
\value{
a list that includes the draws of household-related objects,taking into account the sick parameters and the distribution of coverage, the optimal insurance choice, the amount of oop, and the cost to the insurance company
}
\description{
Compute the counterfactual draws (preference parameters independent). This is different from household_draw_theta_kappa_Rdraw because the draws of thetabar
is only conducted once, and there is no integration over thetabar. Also, the insurance bundles are not one-person deviation from the observed bundle, but
all possible bundles.
}
\examples{
constant_f = function(x) x
counterfactual_household_draw_theta_kappa_Rdraw(3, sample_data_and_parameter$param, n_draw_halton = 1000, n_draw_gauss = 10, sick_parameters = sick_parameters_sample, xi_parameters = xi_parameters_sample, u_lowerbar = -10, policy_mat_hh=policy_mat[[3]], seed_number=1, constraint_function=constant_f, within_hh_heterogeneity = list(omega=TRUE, gamma=TRUE, delta=TRUE, theta_bar=TRUE))


}
