% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{CRRA_fun_Gauss}
\alias{CRRA_fun_Gauss}
\title{Compute the expectation of a CRRA function with Gaussian random variable}
\usage{
CRRA_fun_Gauss(log_R, mu, sigma, n_draw = 50)
}
\arguments{
\item{log_R}{A constant.}

\item{mu}{The mean of the truncated normal random variable}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}

\item{n_draw}{The number of points used to approximate the integral}
}
\value{
A numerical value.
}
\description{
Compute the expectation of a CRRA function with Gaussian random variable
}
\examples{
log_R = 1; mu = 1; sigma = 1; n_draw = 50;
CRRA_fun_Gauss(log_R, mu, sigma, n_draw)
}
