% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{d_mean_E_XW}
\alias{d_mean_E_XW}
\title{Compute the derivative of the expectation of X^y, where y is a truncated normal random variable}
\usage{
d_mean_E_XW(mu, sigma, X)
}
\arguments{
\item{mu}{The mean of the truncated normal random variable}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}

\item{X}{this is a constant}
}
\value{
A list with two values, derivative for the mean and the derivative for the standard deviation
}
\description{
Compute the derivative of the expectation of X^y, where y is a truncated normal random variable
}
\examples{
d_mean_E_XW(1, 1, 1)
}
