% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{d_truncated_normal_variance}
\alias{d_truncated_normal_variance}
\title{Compute the derivative of the variance of truncated normal random variable with respect to its mean and variance}
\usage{
d_truncated_normal_variance(mu, sigma)
}
\arguments{
\item{mu}{The mean of the truncated normal random variable}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}
}
\value{
a list of 2 derivatives, one w.r.t to the mean and one w.r.t to standard deviation
}
\description{
Compute the derivative of the variance of truncated normal random variable with respect to its mean and variance
}
\examples{
d_truncated_normal_variance(1, 1)
}
