% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_hh_list}
\alias{data_hh_list}
\title{Cleaned data on household enrollment and medical expenditure}
\format{
\subsection{\code{data_hh_list}}{

A list, each list element is the data frame for a household.
\describe{
\item{M_expense}{medical expense out of pocket}
\item{tot_cost}{Total medical expenditure (not equal to M_expense only if proportion of coverage is recorded)}
\item{Vol_sts}{Whether a member is voluntarily insured}
\item{Com_sts}{Whether a member is compulsorily insured}
\item{Bef_sts}{Whether a member has free insurance}
}
}
}
\source{
raw_data
}
\usage{
data_hh_list
}
\description{
Cleaned data on household enrollment and medical expenditure
}
\keyword{datasets}
