% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyenrollment-package.R
\docType{package}
\name{familyenrollment-package}
\alias{familyenrollment}
\alias{familyenrollment-package}
\title{familyenrollment: Estimation Procedure for Household Bundling in Health Insurance}
\description{
This package documents all data processing functions and the estimation functions used in "Household Bundling in Health Insurance".
}
\author{
\strong{Maintainer}: Anh Nguyen \email{anhnguyen@cmu.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
