% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{grad_llh_sick}
\alias{grad_llh_sick}
\title{Compute the gradient of LLH of a household member getting sick.}
\usage{
grad_llh_sick(beta, data)
}
\arguments{
\item{beta}{A vector of length var_ind(data)}

\item{data}{A dataframe of a household or multiple households}
}
\value{
a vector, each is a partial derivative w.r.t beta.
}
\description{
Compute the gradient of LLH of a household member getting sick.
}
\examples{
grad_llh_sick(runif(ncol(var_ind(sample_data_and_parameter[[2]]$data))),sample_data_and_parameter[[2]]$data)


}
