% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{grad_llh_xi}
\alias{grad_llh_xi}
\title{Compute the gradient of LLH of a household member getting coverage.}
\usage{
grad_llh_xi(beta, data)
}
\arguments{
\item{beta}{A vector of length var_ind(data)}

\item{data}{A dataframe of a household or multiple households}
}
\value{
a vector, each is a partial derivative w.r.t beta.
}
\description{
Compute the gradient of LLH of a household member getting coverage.
}
\examples{
temp_data = data_hh_list[[34592]]; temp_data$Year = 2008
grad_llh_xi(runif(2 * ncol(var_ind(temp_data))),temp_data)


}
