% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{identify_theta_analytical}
\alias{identify_theta_analytical}
\title{Compute the LLH of medical expenditure.}
\usage{
identify_theta_analytical(
  index,
  param,
  derivative = FALSE,
  numerical_derivative = NA,
  option_derivative = "beta_theta_ind"
)
}
\arguments{
\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}

\item{data_set}{A list produced by household_draw_theta_kappa_Rdraw}

\item{n_draw_halton}{number of halton draws}
}
\value{
a list. The first element is the LLH, the second element is the derivative with respect to beta_theta, beta_theta_ind, and sigma_thetabar
}
\description{
Compute the LLH of medical expenditure.
}
\examples{
identify_theta_analytical(1, transform_param(init_param, return_index=TRUE)[[1]])


}
