% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Income_net_premium}
\alias{Income_net_premium}
\title{Constructed premium vector for each household}
\format{
\subsection{\code{Income_net_premium}}{

A vector
\describe{
\item{Element 1}{Income if no member is insured}
\item{Element 2}{Income if 1 member is insured }
\item{Element 3}{Income if 2 members are insured}
}
}
}
\source{
raw_data
}
\usage{
Income_net_premium
}
\description{
Constructed premium vector for each household
}
\keyword{datasets}
