% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{llh_sick}
\alias{llh_sick}
\title{Compute the LLH of a household member getting sick.}
\usage{
llh_sick(beta, data)
}
\arguments{
\item{beta}{A vector of length var_ind(data)}

\item{data}{A dataframe of a household or multiple households}
}
\value{
a numerical value for LLH.
}
\description{
Compute the LLH of a household member getting sick.
}
\examples{
llh_sick(runif(ncol(var_ind(sample_data_and_parameter[[2]]$data))),sample_data_and_parameter[[2]]$data)

}
