% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{llh_xi}
\alias{llh_xi}
\title{Compute the LLH of a household member getting coverage}
\usage{
llh_xi(beta, data)
}
\arguments{
\item{beta}{A vector of length var_ind(data)}

\item{data}{A dataframe of a household or multiple households}
}
\value{
a numerical value for LLH.
}
\description{
Compute the LLH of a household member getting coverage
}
\examples{
temp_data = data_hh_list[[34592]]; temp_data$Year = 2008
llh_xi(runif(2 * ncol(var_ind(temp_data))),temp_data)

}
