% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{moment_ineligible_hh}
\alias{moment_ineligible_hh}
\title{Compute the moments of ineligible households}
\usage{
moment_ineligible_hh(data_set, param)
}
\arguments{
\item{data_set}{This is a list that contains all pre-computed elements of each household .}

\item{param}{This is a list, each element is a coefficient (i.e., a vector) of a preference parameter (beta_gamma, beta_omega, beta_delta, beta_theta, beta_theta_ind) and the standard deviations of the unobserved heterogeneity (sigma_gamma, sigma_omega, sigma_delta, sigma_theta, sigma_thetabar)}
}
\value{
a list. The first element is the expected value of m|Y,k. The second element is the expected value of m^2|Y,k. The third and the fourth elements are their respective derivatives with respect to the preference parameters.
}
\description{
Compute the moments of ineligible households
}
\examples{
mini_data = household_draw_theta_kappa_Rdraw(2, sample_data_and_parameter$param, 1000, 10, sick_parameters_sample, xi_parameters_sample, short=FALSE);
moment_ineligible_hh(mini_data, sample_data_and_parameter$param)


}
