% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy_mat}
\alias{policy_mat}
\title{Constructed coinsurance matrices}
\format{
\subsection{\code{policy_mat}}{

A list, with three elements, each is a matrix of HHsize columns and 10 rows .
\describe{
\item{List 1}{The coinsurance rate of each linear segment}
\item{List 2}{The threshold at which coinsurance rates change}
\item{List 3}{The out of pocket costs at the nearest threshold}
}
}
}
\source{
raw_data
}
\usage{
policy_mat
}
\description{
Constructed coinsurance matrices
}
\keyword{datasets}
