% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data_and_parameter}
\alias{sample_data_and_parameter}
\title{Sample data used to test moment of eligible households}
\format{
\subsection{\code{sample_data_and_parameter}}{

A list with two elements
\describe{
\item{data}{A list with different vectors characterizing a household}
\item{param}{A list with different vectors characterizing parameters}
}
}
}
\source{
sample_data_and_parameter
}
\usage{
sample_data_and_parameter
}
\description{
A subset of households' constructed data after estimation of the preference parameters and the unconditional health shocks
}
\keyword{datasets}
