% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{second_taylor_CARA}
\alias{second_taylor_CARA}
\title{Compute the expectation of a CARA expected utility function}
\usage{
second_taylor_CARA(a, b, mu, sigma)
}
\arguments{
\item{a}{A constant, which is the mean of (direct) utility.}

\item{b}{A constant, which is the variance of (indirect) utility.}

\item{mu}{The mean of the truncated normal random variable}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}
}
\value{
A numerical value.
}
\description{
Compute the expectation of a CARA expected utility function
}
\examples{
second_taylor_CARA(1,1,1,1)
}
