% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{second_taylor_CARA_test}
\alias{second_taylor_CARA_test}
\title{Compute the expectation of a CARA expected utility function using numerical simulations}
\usage{
second_taylor_CARA_test(a, b, mu, sigma)
}
\arguments{
\item{a}{A constant, which is the mean of (direct) utility.}

\item{b}{A constant, which is the variance of (indirect) utility.}

\item{mu}{The mean of the truncated normal random variable.}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}
}
\value{
A list. The first element is the numerical integration, the second element is the computation based on the second-order Taylor approximation.
}
\description{
Compute the expectation of a CARA expected utility function using numerical simulations
}
\examples{
second_taylor_CARA_test(1,1,1,1)
}
