% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sick_parameters_sample}
\alias{sick_parameters_sample}
\title{Example of estimates for the parameters characterizing whether a household member is going to be sick}
\format{
\subsection{\code{sick_parameters_sample}}{

A list produced by optim
}
}
\source{
raw_data
}
\usage{
sick_parameters_sample
}
\description{
Example of estimates for the parameters characterizing whether a household member is going to be sick
}
\keyword{datasets}
