% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{transform_param}
\alias{transform_param}
\title{Construct parameters list.}
\usage{
transform_param(param_trial, return_index = FALSE, init = FALSE)
}
\arguments{
\item{param_trial}{A vector}

\item{return_index}{LOGICAL, whether the position of the parameters is also returned.}

\item{init}{LOGICAL, whether the initial value needs to supplied}
}
\value{
if return_index is TRUE, return a list, where the first element is a list of parameters (beta_theta, beta_delta, etc..) and the second element is the position of the parameters from the original vector. If return_index is FALSE, only the list of parameters is returned. If init==TRUE, return a vector of 0 with dimension equal to the dimension required.
}
\description{
Construct parameters list.
}
\examples{
transform_param(return_index=TRUE, init=TRUE)


}
