% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{trunc_norm}
\alias{trunc_norm}
\title{Compute the mean and variance of a truncated from below normal distribution}
\usage{
trunc_norm(mu, sigma, a = 0)
}
\arguments{
\item{mu}{mean of the (untruncated) normal variable}

\item{sigma}{standard deviation of the (untruncated) normal variable}

\item{a}{is the truncation point. DEFAULT to 0.}
}
\value{
a list of two values, mean and variance.
}
\description{
Compute the mean and variance of a truncated from below normal distribution
}
\examples{
trunc_norm(1,1,2)
}
