% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{truncated_normal_variance}
\alias{truncated_normal_variance}
\title{Compute the variance of truncated normal random variable}
\usage{
truncated_normal_variance(mu, sigma)
}
\arguments{
\item{mu}{The mean of the truncated normal random variable}

\item{sigma}{The standard deviation of the truncated normal random variable (truncated at 0)}
}
\value{
variance of truncated normal distribution at 0
}
\description{
Compute the variance of truncated normal random variable
}
\examples{
truncated_normal_variance(1, 1)
}
