% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{U}
\alias{U}
\title{Compute direct utility function}
\usage{
U(input, income_effect = TRUE)
}
\arguments{
\item{input}{is a list of all input components.}
}
\value{
a vector of utility realization
}
\description{
Compute direct utility function
}
\examples{
U(list(R_draw = runif(10), theta_draw = cbind(runif(10)/10, runif(10)/10), delta = c(0.5, 0.5), omega=0.1, gamma = c(0.5, 0.5), kappa_draw = matrix(1, nrow=10, ncol=2), HHsize = 2))

}
