% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{uniroot_usr}
\alias{uniroot_usr}
\title{Finding the root of a function, used to compute willingness to pay.}
\usage{
uniroot_usr(f, interval_init)
}
\arguments{
\item{f}{A function}

\item{interval_init}{is an interval}
}
\value{
return an object produced by uniroot (if a root can be found)
}
\description{
Finding the root of a function, used to compute willingness to pay.
}
\examples{
uniroot_usr(function(x) x^2 - 2, c(1, 2))


}
