% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_inc}
\alias{unit_inc}
\title{unit of income in estimation}
\format{
\subsection{\code{unit_inc}}{

a numerical value
}
}
\source{
raw_data
}
\usage{
unit_inc
}
\description{
unit of income in estimation
}
\keyword{datasets}
