% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{var_hh}
\alias{var_hh}
\title{Compute the matrix representing household characteristics of a household.}
\usage{
var_hh(data_mini)
}
\arguments{
\item{data_mini}{A dataframe of a household}
}
\value{
a vector of X_hh.
}
\description{
Compute the matrix representing household characteristics of a household.
}
\examples{
var_hh(data_hh_list[[1]])

}
