% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_functions.R
\name{var_ind}
\alias{var_ind}
\title{Compute the matrix representing individual characteristics of a household.}
\usage{
var_ind(data_mini)
}
\arguments{
\item{data_mini}{A dataframe of a household's characteristics}
}
\value{
a matrix with nrow being equal to the number of household members.
}
\description{
Compute the matrix representing individual characteristics of a household.
}
\examples{
var_ind(data_hh_list[[1]])

}
