% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xi_parameters_sample}
\alias{xi_parameters_sample}
\title{Example of estimates for the parameters characterizing whether a household member is going to be covered under SHI}
\format{
\subsection{\code{xi_parameters_sample}}{

A list produced by optim
}
}
\source{
raw_data
}
\usage{
xi_parameters_sample
}
\description{
Example of estimates for the parameters characterizing whether a household member is going to be covered under SHI
}
\keyword{datasets}
